Refactor community comment activity, Add transition animation when community comment activity finished
