+++ client/android/app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PetManagement">
-        <activity android:name=".ui.community.CommunityActivity"></activity>
+        <activity android:name=".ui.community.comment.CommunityCommentActivity"></activity>
         <activity android:name=".ui.welcomePage.WelcomePageActivity" />
         <activity android:name=".ui.login.LoginActivity">
             <intent-filter>
