+++ client/android/app/src/main/java/com/sju18001/petmanagement/ui/community/comment/CommunityCommentActivity.kt
@@ -1,0 +1,30 @@
+package com.sju18001.petmanagement.ui.community.comment
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.sju18001.petmanagement.R
+import com.sju18001.petmanagement.databinding.ActivityCommunityCommentBinding
+
+class CommunityCommentActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityCommunityCommentBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityCommunityCommentBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        supportActionBar?.hide()
+
+        // 프래그먼트 실행
+        supportFragmentManager
+            .beginTransaction()
+            .add(R.id.community_activity_fragment_container, CommunityCommentFragment())
+            .commit()
+    }
+
+    override fun finish() {
+        super.finish()
+        overridePendingTransition(R.anim.enter_from_left, R.anim.exit_to_right)
+    }
+}
