+++ client/android/app/src/main/java/com/sju18001/petmanagement/ui/community/comment/CommunityCommentListAdapter.kt
@@ -1,4 +1,4 @@
-package com.sju18001.petmanagement.ui.community
+package com.sju18001.petmanagement.ui.community.comment
 
 import android.view.LayoutInflater
 import android.view.View
