+++ client/android/app/src/main/res/layout/fragment_community_comment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.community.CommunityCommentFragment">
+    tools:context=".ui.community.comment.CommunityCommentFragment">
 
     <RelativeLayout
         android:id="@+id/back_button_layout"
